Moving Utils and adding locationIfComplex
